AgencyとMetroのソースの構造を合わせる
*BarChart.vueにはチャートデザインと動作
/card/*.vueにはデータを置く
