+++ components/cards/AgencyCard.vue
@@ -4,6 +4,8 @@
       :title="$t('都庁来庁者数の推移')"
       :title-id="'agency'"
       :chart-id="'agency'"
+      :chart-data="agencyData"
+      :date="agencyData.date"
       :url="''"
       :unit="$t('人')"
     />
@@ -40,11 +42,26 @@
 </i18n>
 
 <script>
+import agencyData from '@/data/agency.json'
 import AgencyBarChart from '@/components/AgencyBarChart.vue'
 
 export default {
   components: {
     AgencyBarChart
+  },
+  data() {
+    const agencies = [
+      this.$t('第一庁舎計'),
+      this.$t('第二庁舎計'),
+      this.$t('議事堂計')
+    ]
+    agencyData.datasets.map(dataset => {
+      dataset.label = this.$t(dataset.label)
+    })
+    return {
+      agencyData,
+      agencies
+    }
   }
 }
 </script>
